refactor initial assessment - removes first seizure - refactors toggles
