+++ epilepsy12/view_folder/initial_assessment_views.py
@@ -143,26 +143,22 @@ def general_paediatrics_referral_made(request, initial_assessment_id):
     same partial.
     """
 
-    initial_assessment = InitialAssessment.objects.get(
-        pk=initial_assessment_id)
-
-    general_paediatrics_referral_made = not initial_assessment.general_paediatrics_referral_made
-
-    if general_paediatrics_referral_made:
-        InitialAssessment.objects.filter(
-            pk=initial_assessment_id).update(
-                general_paediatrics_referral_made=True,
-                updated_at=timezone.now(),
-                updated_by=request.user
+    if request.htmx.trigger_name == 'button-true':
+        InitialAssessment.objects.filter(pk=initial_assessment_id).update(
+            general_paediatrics_referral_made=True,
+            updated_at=timezone.now(),
+            updated_by=request.user
+        )
+    elif request.htmx.trigger_name == 'button-false':
+        InitialAssessment.objects.filter(pk=initial_assessment_id).update(
+            general_paediatrics_referral_made=False,
+            date_of_referral_to_general_paediatrics=None,
+            updated_at=timezone.now(),
+            updated_by=request.user
         )
     else:
-        InitialAssessment.objects.filter(
-            pk=initial_assessment_id).update(
-                general_paediatrics_referral_made=False,
-                date_of_referral_to_general_paediatrics=None,
-                updated_at=timezone.now(),
-                updated_by=request.user
-        )
+        print("Some mistake happened")
+        # TODO need to handle this
 
     initial_assessment = InitialAssessment.objects.get(
         pk=initial_assessment_id)
@@ -226,120 +222,47 @@ def date_of_referral_to_general_paediatrics(request, initial_assessment_id):
     return response
 
 
-def when_the_first_epileptic_episode_occurred(request, initial_assessment_id):
+# @login_required
+# def has_description_of_the_episode_or_episodes_been_gathered(request, initial_assessment_id):
+#     """
+#     POST request multiple choice toggle in when_the_first_epileptic_episode_occurred partial
+#     """
 
-    when_the_first_epileptic_episode_occurred = request.POST.get(
-        'when_the_first_epileptic_episode_occurred')
+#     if request.htmx.trigger_name == 'button-true':
+#         InitialAssessment.objects.filter(pk=initial_assessment_id).update(
+#             has_description_of_the_episode_or_episodes_been_gathered=True,
+#             updated_at=timezone.now(),
+#             updated_by=request.user
+#         )
+#     elif request.htmx.trigger_name == 'button-false':
+#         InitialAssessment.objects.filter(pk=initial_assessment_id).update(
+#             has_description_of_the_episode_or_episodes_been_gathered=False,
+#             updated_at=timezone.now(),
+#             updated_by=request.user
+#         )
+#     else:
+#         print("Some mistake happened")
+#         # TODO need to handle this
 
-    if when_the_first_epileptic_episode_occurred:
-        new_date = datetime.strptime(
-            when_the_first_epileptic_episode_occurred, "%Y-%m-%d").date()
-        # TODO validation goes here
+#     initial_assessment = InitialAssessment.objects.get(
+#         pk=initial_assessment_id)
 
-        try:
-            InitialAssessment.objects.filter(pk=initial_assessment_id).update(
-                when_the_first_epileptic_episode_occurred=new_date,
-                updated_at=timezone.now(),
-                updated_by=request.user
-            )
-        except Exception as error:
-            message = error
+#     context = {
+#         "initial_assessment": initial_assessment,
+#         "when_the_first_epileptic_episode_occurred_confidence_selection": DATE_ACCURACY,
+#         "diagnostic_status_selection": DIAGNOSTIC_STATUS,
+#         "episode_definition_selection": EPISODE_DEFINITION,
+#     }
 
-    initial_assessment = InitialAssessment.objects.get(
-        pk=initial_assessment_id)
+#     response = render(
+#         request=request, template_name="epilepsy12/partials/initial_assessment/when_the_first_epileptic_episode_occurred.html", context=context)
 
-    context = {
-        'initial_assessment_id': initial_assessment_id,
-        'initial_assessment': initial_assessment,
-        "chronicity_selection": CHRONICITY,
-        "when_the_first_epileptic_episode_occurred_confidence_selection": DATE_ACCURACY,
-        "diagnostic_status_selection": DIAGNOSTIC_STATUS,
-        "episode_definition_selection": EPISODE_DEFINITION,
-    }
-
-    test_fields_update_audit_progress(initial_assessment)
-
-    response = render(
-        request=request, template_name="epilepsy12/partials/initial_assessment/when_the_first_epileptic_episode_occurred.html", context=context)
-
-    # trigger a GET request from the steps template
-    trigger_client_event(
-        response=response,
-        name="registration_active",
-        params={})  # reloads the form to show the active steps
-    return response
-
-
-def when_the_first_epileptic_episode_occurred_confidence(request, initial_assessment_id):
-    """
-    HTMX callback from when_the_first_epileptic_episode_occurred
-    """
-    when_the_first_epileptic_episode_occurred_confidence = request.htmx.trigger_name
-
-    if when_the_first_epileptic_episode_occurred_confidence:
-        try:
-            InitialAssessment.objects.filter(pk=initial_assessment_id).update(
-                when_the_first_epileptic_episode_occurred_confidence=when_the_first_epileptic_episode_occurred_confidence,
-                updated_at=timezone.now(),
-                updated_by=request.user
-            )
-        except Exception as error:
-            message = error
-
-    initial_assessment = InitialAssessment.objects.get(
-        pk=initial_assessment_id)
-
-    context = {
-        'initial_assessment_id': initial_assessment_id,
-        'initial_assessment': initial_assessment,
-        "chronicity_selection": CHRONICITY,
-        "when_the_first_epileptic_episode_occurred_confidence_selection": DATE_ACCURACY,
-        "diagnostic_status_selection": DIAGNOSTIC_STATUS,
-        "episode_definition_selection": EPISODE_DEFINITION,
-    }
-
-    response = render(
-        request=request, template_name="epilepsy12/partials/initial_assessment/when_the_first_epileptic_episode_occurred.html", context=context)
-
-    # trigger a GET request from the steps template
-    trigger_client_event(
-        response=response,
-        name="registration_active",
-        params={})  # reloads the form to show the active steps
-    return response
-
-
-def has_description_of_the_episode_or_episodes_been_gathered(request, initial_assessment_id):
-
-    initial_assessment = InitialAssessment.objects.get(
-        pk=initial_assessment_id)
-    has_description_of_the_episode_or_episodes_been_gathered = not initial_assessment.has_description_of_the_episode_or_episodes_been_gathered
-
-    try:
-        InitialAssessment.objects.filter(pk=initial_assessment_id).update(
-            has_description_of_the_episode_or_episodes_been_gathered=has_description_of_the_episode_or_episodes_been_gathered)
-    except Exception as error:
-        return HttpResponse(error)
-
-    new_initial_assessment = InitialAssessment.objects.get(
-        pk=initial_assessment_id)
-
-    context = {
-        "initial_assessment": new_initial_assessment,
-        "when_the_first_epileptic_episode_occurred_confidence_selection": DATE_ACCURACY,
-        "diagnostic_status_selection": DIAGNOSTIC_STATUS,
-        "episode_definition_selection": EPISODE_DEFINITION,
-    }
-
-    response = render(
-        request=request, template_name="epilepsy12/partials/initial_assessment/when_the_first_epileptic_episode_occurred.html", context=context)
-
-    # trigger a GET request from the steps template
-    trigger_client_event(
-        response=response,
-        name="registration_active",
-        params={})  # reloads the form to show the active steps
-    return response
+#     # trigger a GET request from the steps template
+#     trigger_client_event(
+#         response=response,
+#         name="registration_active",
+#         params={})  # reloads the form to show the active steps
+#     return response
 
 
 def has_number_of_episodes_since_the_first_been_documented(request, initial_assessment_id):
@@ -373,7 +296,7 @@ def has_number_of_episodes_since_the_first_been_documented(request, initial_asse
     }
 
     response = render(
-        request=request, template_name="epilepsy12/partials/initial_assessment/has_number_of_episodes_since_the_first_been_documented.html", context=context)
+        request=request, template_name="epilepsy12/partials/initial_assessment/when_the_first_epileptic_episode_occurred.html", context=context)
 
     # trigger a GET request from the steps template
     trigger_client_event(
@@ -414,7 +337,7 @@ def general_examination_performed(request, initial_assessment_id):
     }
 
     response = render(
-        request=request, template_name="epilepsy12/partials/initial_assessment/general_examination_performed.html", context=context)
+        request=request, template_name="epilepsy12/partials/initial_assessment/when_the_first_epileptic_episode_occurred.html", context=context)
 
     # trigger a GET request from the steps template
     trigger_client_event(
@@ -455,7 +378,7 @@ def neurological_examination_performed(request, initial_assessment_id):
     }
 
     response = render(
-        request=request, template_name="epilepsy12/partials/initial_assessment/neurological_examination_performed.html", context=context)
+        request=request, template_name="epilepsy12/partials/initial_assessment/when_the_first_epileptic_episode_occurred.html", context=context)
 
     # trigger a GET request from the steps template
     trigger_client_event(
@@ -496,7 +419,7 @@ def developmental_learning_or_schooling_problems(request, initial_assessment_id)
     }
 
     response = render(
-        request=request, template_name="epilepsy12/partials/initial_assessment/developmental_learning_or_schooling_problems.html", context=context)
+        request=request, template_name="epilepsy12/partials/initial_assessment/when_the_first_epileptic_episode_occurred.html", context=context)
 
     # trigger a GET request from the steps template
     trigger_client_event(
@@ -537,7 +460,7 @@ def behavioural_or_emotional_problems(request, initial_assessment_id):
     }
 
     response = render(
-        request=request, template_name="epilepsy12/partials/initial_assessment/behavioural_or_emotional_problems.html", context=context)
+        request=request, template_name="epilepsy12/partials/initial_assessment/when_the_first_epileptic_episode_occurred.html", context=context)
 
     # trigger a GET request from the steps template
     trigger_client_event(
@@ -560,45 +483,6 @@ def diagnostic_status(request, initial_assessment_id):
                 updated_at=timezone.now(),
                 updated_by=request.user
         )
-    except Exception as error:
-        print(error)
-        return HttpResponse(error)
-
-    initial_assessment = InitialAssessment.objects.get(
-        pk=initial_assessment_id)
-
-    context = {
-        "initial_assessment": initial_assessment,
-        "when_the_first_epileptic_episode_occurred_confidence_selection": DATE_ACCURACY,
-        "diagnostic_status_selection": DIAGNOSTIC_STATUS,
-        "episode_definition_selection": EPISODE_DEFINITION,
-    }
-
-    test_fields_update_audit_progress(initial_assessment)
-
-    response = render(
-        request=request, template_name="epilepsy12/partials/initial_assessment/when_the_first_epileptic_episode_occurred.html", context=context)
-
-# trigger a GET request from the steps template
-    trigger_client_event(
-        response=response,
-        name="registration_active",
-        params={})  # reloads the form to show the active steps
-    return response
-
-
-def episode_definition(request, initial_assessment_id):
-
-    episode_definition = request.POST.get(
-        'episode_definition')
-    # validation here TODO
-
-    try:
-        InitialAssessment.objects.filter(
-            pk=initial_assessment_id).update(
-                episode_definition=episode_definition,
-                updated_at=timezone.now(),
-                updated_by=request.user)
     except Exception as error:
         print(error)
         return HttpResponse(error)
@@ -653,7 +537,6 @@ def completed_fields(model_instance):
     developmental_learning_or_schooling_problems
     behavioural_or_emotional_problems
     diagnostic_status
-    episode_definition
     """
     fields = model_instance._meta.get_fields()
     counter = 0
@@ -662,7 +545,6 @@ def completed_fields(model_instance):
             getattr(model_instance, field.name) is not None
             and field.name not in ['id', 'registration', 'created_at', 'updated_at', 'created_by', 'updated_by']
         ):
-            print(field.name)
             counter += 1
     return counter
