+++ templates/epilepsy12/forms/initial_assessment_form.html
@@ -25,33 +25,11 @@
   </div>
 
   <h5 class='header'>
-    The first episode
+    Information gathered at the first paediatric assessment
   </h5>
 
   <div id='when_the_first_epileptic_episode_occurred'>
     {% include 'epilepsy12/partials/initial_assessment/when_the_first_epileptic_episode_occurred.html' with initial_assessment=initial_assessment when_the_first_epileptic_episode_occurred_confidence_selection=when_the_first_epileptic_episode_occurred_confidence_selection diagnostic_status_selection=diagnostic_status_selection %}
   </div>
-  
-  <div id='has_number_of_episodes_since_the_first_been_documented'>
-    {% include 'epilepsy12/partials/initial_assessment/has_number_of_episodes_since_the_first_been_documented.html' with initial_assessment=initial_assessment  %}
-  </div>
-  
-  <div id='general_examination_performed'>
-    {% include 'epilepsy12/partials/initial_assessment/general_examination_performed.html' with initial_assessment=initial_assessment  %}
-  </div>
-  
-  <div id='neurological_examination_performed'>
-    {% include 'epilepsy12/partials/initial_assessment/neurological_examination_performed.html' with initial_assessment=initial_assessment  %}
-  </div>
-  
-  <div id='developmental_learning_or_schooling_problems'>
-    {% include 'epilepsy12/partials/initial_assessment/developmental_learning_or_schooling_problems.html' with initial_assessment=initial_assessment  %}
-  </div>
-  
-  <div id='behavioural_or_emotional_problems'>
-    {% include 'epilepsy12/partials/initial_assessment/behavioural_or_emotional_problems.html' with initial_assessment=initial_assessment  %}
-  </div>
-
-
   
 </form>
