--- templates/epilepsy12/partials/initial_assessment/general_examination_performed.html
@@ -1,4 +1,0 @@
-<div class='field' >
-    {% url 'general_examination_performed' initial_assessment_id=initial_assessment.pk as hx_post %}
-    {% include 'epilepsy12/partials/page_elements/toggle_button.html' with hx_post=hx_post hx_target='#general_examination_performed' hx_trigger='click' hx_swap='innerHTML' test_positive=initial_assessment.general_examination_performed label='Has a general examination been performed?' data_position='top left' %}
-</div>
