+++ templates/epilepsy12/partials/page_elements/select.html
@@ -15,23 +15,24 @@
         The chosen value is stored as the name of the button and accessed in the view as request.htmx.trigger_name
         Which choices are selected from is obvious because the endpoint posted to is specific to the choices.
     {% endcomment %}
-    <label class="rcpch_plain_label">
-        {{label}}
-    </label>
     <div class="field" hx-headers='{"X-CSRFToken": "{{ csrf_token }}"}'>
-
-        <span>
-            {% if test_positive %}
-                <i class="rcpch_pink check circle outline icon"></i>
-            {% else %}
-                <span data-tooltip="Uncompleted field. This must be scored to complete the record." data-inverted="" data-position="top left">
-                    <i class="rcpch_light_blue dot circle outline icon"></i>
-                </span>
-            {% endif %}
+        
+            <label class="rcpch_plain_label">
+                {{label}}
+            </label>
 
             <div 
-                class="ui rcpch_light_blue selection dropdown"
+                class="ui rcpch_light_blue fluid selection dropdown"
             >
+            
+                    {% if test_positive %}
+                        <i class="rcpch_pink check circle outline icon"></i>
+                    {% else %}
+                        <span data-tooltip="Uncompleted field. This must be scored to complete the record." data-inverted="" data-position="top left">
+                            <i class="rcpch_light_blue dot circle outline icon"></i>
+                        </span>
+                    {% endif %}
+
                     <i class="dropdown icon"></i>
                     <input type="hidden" 
                         name={{hx_name}}
@@ -49,7 +50,7 @@
                     </div>
             </div>
 
-        </span>
+        
 
     </div>
